+++ pyriemann_qiskit/utils/__init__.py
@@ -1,10 +1,7 @@
 from . import (
-    anderson_optimizer,
-    distance,
-    docplex,
     filtering,
     hyper_params_factory,
-    mean,
+    math,
     preprocessing,
     quantum_provider,
     transfer,
@@ -13,15 +10,12 @@ from . import (
 from .math import union_of_diff
 
 __all__ = [
-    "anderson_optimizer",
     "hyper_params_factory",
     "filtering",
+    "math",
     "preprocessing",
     "quantum_provider",
     "union_of_diff",
-    "docplex",
-    "distance",
-    "mean",
     "transfer",
     "utils",
 ]