+++ tests/test_anderson_optimizer.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from pyriemann_qiskit.utils.anderson_optimizer import AndersonAccelerationOptimizer
+from pyriemann_qiskit.optimization.anderson_optimizer import AndersonAccelerationOptimizer
 
 
 def quadratic(x):