+++ tests/test_utils_distance.py
@@ -7,11 +7,11 @@ from sklearn.pipeline import make_pipeline
 
 from pyriemann_qiskit.classification import QuanticMDM
 from pyriemann_qiskit.datasets import get_mne_sample
-from pyriemann_qiskit.utils.distance import (
+from pyriemann_qiskit.optimization.distance import (
     qdistance_logeuclid_to_convex_hull,
     weights_logeuclid_to_convex_hull,
 )
-from pyriemann_qiskit.utils.docplex import (  # NaiveQAOAOptimizer,; QAOACVOptimizer,
+from pyriemann_qiskit.optimization.docplex import (  # NaiveQAOAOptimizer,; QAOACVOptimizer,
     ClassicalOptimizer,
 )